# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SRI - Release Go Live Control API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /escala:
    get:
      operationId: listEscala
      tags: [escala]
      summary: List on-call schedule entries
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of schedule entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Escala"
    post:
      operationId: createEscala
      tags: [escala]
      summary: Create a schedule entry (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EscalaInput"
      responses:
        "201":
          description: Created entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Escala"
  /escala/{id}:
    patch:
      operationId: updateEscala
      tags: [escala]
      summary: Update a schedule entry (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EscalaUpdate"
      responses:
        "200":
          description: Updated entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Escala"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteEscala
      tags: [escala]
      summary: Delete a schedule entry (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /contatos:
    get:
      operationId: listContatos
      tags: [contatos]
      summary: List contacts
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contato"
    post:
      operationId: createContato
      tags: [contatos]
      summary: Create a contact (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContatoInput"
      responses:
        "201":
          description: Created contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contato"
  /contatos/{id}:
    patch:
      operationId: updateContato
      tags: [contatos]
      summary: Update a contact (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContatoUpdate"
      responses:
        "200":
          description: Updated contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contato"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteContato
      tags: [contatos]
      summary: Delete a contact (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Consolidated dashboard counts
      parameters:
        - name: prioridade
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/matrix:
    get:
      operationId: getDashboardMatrix
      tags: [dashboard]
      summary: Site x Status matrix with totals and percentages
      parameters:
        - name: prioridade
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Matrix
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardMatrix"
  /scenarios/{id}/evidencias:
    get:
      operationId: listEvidencias
      tags: [evidencias]
      summary: List evidence files for a scenario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of evidences
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Evidencia"
    post:
      operationId: createEvidencia
      tags: [evidencias]
      summary: Register an uploaded evidence file (uploaders only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvidenciaInput"
      responses:
        "201":
          description: Created evidence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evidencia"
        "404":
          description: Scenario not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /evidencias/{id}:
    delete:
      operationId: deleteEvidencia
      tags: [evidencias]
      summary: Delete an evidence file (uploaders only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Escala:
      type: object
      required: [id, pessoa]
      properties:
        id:
          type: integer
        pessoa:
          type: string
        empresa:
          type: ["string", "null"]
        papel:
          type: ["string", "null"]
        dia:
          type: ["string", "null"]
        horaInicio:
          type: ["string", "null"]
        horaFim:
          type: ["string", "null"]
        createdAt:
          type: ["string", "null"]
    EscalaInput:
      type: object
      required: [pessoa]
      properties:
        pessoa:
          type: string
          minLength: 1
        empresa:
          type: string
        papel:
          type: string
        dia:
          type: string
        horaInicio:
          type: string
        horaFim:
          type: string
    EscalaUpdate:
      type: object
      properties:
        pessoa:
          type: string
        empresa:
          type: string
        papel:
          type: string
        dia:
          type: string
        horaInicio:
          type: string
        horaFim:
          type: string
    Contato:
      type: object
      required: [id, nome, empresa]
      properties:
        id:
          type: integer
        nome:
          type: string
        empresa:
          type: string
        contato1:
          type: ["string", "null"]
        contato2:
          type: ["string", "null"]
        localidade:
          type: ["string", "null"]
        papel:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        escalonamento:
          type: ["string", "null"]
        createdAt:
          type: ["string", "null"]
    ContatoInput:
      type: object
      required: [nome, empresa]
      properties:
        nome:
          type: string
          minLength: 1
        empresa:
          type: string
          minLength: 1
        contato1:
          type: string
        contato2:
          type: string
        localidade:
          type: string
        papel:
          type: string
        email:
          type: string
        escalonamento:
          type: string
    ContatoUpdate:
      type: object
      properties:
        nome:
          type: string
        empresa:
          type: string
        contato1:
          type: string
        contato2:
          type: string
        localidade:
          type: string
        papel:
          type: string
        email:
          type: string
        escalonamento:
          type: string
    CountItem:
      type: object
      required: [label, count]
      properties:
        label:
          type: string
        count:
          type: integer
    DashboardSummary:
      type: object
      required: [total, byStatus, bySite, bySistema, byPrioridade]
      properties:
        total:
          type: integer
        byStatus:
          type: array
          items:
            $ref: "#/components/schemas/CountItem"
        bySite:
          type: array
          items:
            $ref: "#/components/schemas/CountItem"
        bySistema:
          type: array
          items:
            $ref: "#/components/schemas/CountItem"
        byPrioridade:
          type: array
          items:
            $ref: "#/components/schemas/CountItem"
    MatrixRow:
      type: object
      required: [site, cells, total]
      properties:
        site:
          type: string
        cells:
          type: array
          items:
            type: integer
        total:
          type: integer
    DashboardMatrix:
      type: object
      required: [statuses, rows, columnTotals, grandTotal, statusPercentages]
      properties:
        statuses:
          type: array
          items:
            type: string
        rows:
          type: array
          items:
            $ref: "#/components/schemas/MatrixRow"
        columnTotals:
          type: array
          items:
            type: integer
        grandTotal:
          type: integer
        statusPercentages:
          type: array
          items:
            type: number
    Evidencia:
      type: object
      required: [id, scenarioId, objectPath, fileName]
      properties:
        id:
          type: integer
        scenarioId:
          type: integer
        objectPath:
          type: string
        fileName:
          type: string
        contentType:
          type: ["string", "null"]
        size:
          type: ["integer", "null"]
        uploadedByEmail:
          type: ["string", "null"]
        createdAt:
          type: ["string", "null"]
    EvidenciaInput:
      type: object
      required: [objectPath, fileName]
      properties:
        objectPath:
          type: string
          minLength: 1
        fileName:
          type: string
          minLength: 1
        contentType:
          type: string
        size:
          type: integer
...